+++ test/pleroma/web/mastodon_api/views/scheduled_activity_view_test.exs
@@ -58,7 +58,8 @@ defmodule Pleroma.Web.MastodonAPI.ScheduledActivityViewTest do
         sensitive: true,
         spoiler_text: "spoiler",
         text: "hi",
-        visibility: "unlisted"
+        visibility: "unlisted",
+        expires_in: nil
       },
       scheduled_at: Utils.to_masto_date(scheduled_activity.scheduled_at)
     }